Move indexer resetter store.
