+++ enterprise/cmd/frontend/internal/executorqueue/queues/codeintel/queue.go
@@ -5,22 +5,20 @@ import (
 
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/frontend/internal/executorqueue/handler"
 	apiclient "github.com/sourcegraph/sourcegraph/enterprise/internal/executor"
-	store "github.com/sourcegraph/sourcegraph/internal/codeintel/stores/dbstore"
+	"github.com/sourcegraph/sourcegraph/internal/codeintel/autoindexing"
 	"github.com/sourcegraph/sourcegraph/internal/codeintel/types"
-	"github.com/sourcegraph/sourcegraph/internal/database"
-	"github.com/sourcegraph/sourcegraph/internal/database/basestore"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
 	"github.com/sourcegraph/sourcegraph/internal/workerutil"
 )
 
-func QueueOptions(db database.DB, accessToken func() string, observationContext *observation.Context) handler.QueueOptions {
+func QueueOptions(autoIndexingSvc *autoindexing.Service, accessToken func() string, observationContext *observation.Context) handler.QueueOptions {
 	recordTransformer := func(ctx context.Context, record workerutil.Record) (apiclient.Job, error) {
 		return transformRecord(record.(types.Index), accessToken())
 	}
 
 	return handler.QueueOptions{
 		Name:              "codeintel",
-		Store:             store.WorkerutilIndexStore(basestore.NewWithHandle(db.Handle()), observationContext),
+		Store:             autoIndexingSvc.WorkerutilStore(),
 		RecordTransformer: recordTransformer,
 	}
 }
