+++ enterprise/cmd/worker/internal/codeintel/janitor_job.go
@@ -13,7 +13,7 @@ import (
 	workerdb "github.com/sourcegraph/sourcegraph/cmd/worker/shared/init/db"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/worker/internal/codeintel/janitor"
 	"github.com/sourcegraph/sourcegraph/enterprise/cmd/worker/internal/executorqueue"
-	"github.com/sourcegraph/sourcegraph/internal/codeintel/stores/dbstore"
+	"github.com/sourcegraph/sourcegraph/internal/codeintel/autoindexing"
 	"github.com/sourcegraph/sourcegraph/internal/codeintel/uploads"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/env"
@@ -58,7 +58,10 @@ func (j *janitorJob) Routines(startupCtx context.Context, logger log.Logger) ([]
 	if err != nil {
 		return nil, err
 	}
+
+	repoUpdater := codeintel.InitRepoUpdaterClient()
 	uploadSvc := uploads.GetService(db, codeIntelDB, gitserverClient)
+	autoindexingSvc := autoindexing.GetService(db, uploadSvc, gitserverClient, repoUpdater)
 
 	dependencyIndexingStore, err := codeintel.InitDependencySyncingStore()
 	if err != nil {
@@ -66,7 +69,7 @@ func (j *janitorJob) Routines(startupCtx context.Context, logger log.Logger) ([]
 	}
 
 	uploadWorkerStore := uploadSvc.WorkerutilStore()
-	indexWorkerStore := dbstore.WorkerutilIndexStore(db, observationContext)
+	indexWorkerStore := autoindexingSvc.WorkerutilStore()
 	metrics := janitor.NewMetrics(observationContext)
 
 	executorMetricsReporter, err := executorqueue.NewMetricReporter(observationContext, "codeintel", indexWorkerStore, janitorConfigInst.MetricsConfig)
